getproto improvement & loadstring fix
loadstring: This was supposed to be fixed in luau but was never fixed. There was a comment left on it, you can view it here: https://github.com/luau-lang/luau/blob/c51743268bcbe5c5af1bd78bcacaefe7f6fe3391/Analysis/src/EmbeddedBuiltinDefinitions.cpp#L51

getproto: doing it this way allows for 2 outcomes based on the third argument. Having it the way it was before would trigger this on the luau-lsp linter: https://imgur.com/a/6o3hKSz

With the improvement: https://imgur.com/a/RcIiYgV